testsuite: follow upstream on QSR operand types and runtime immediates
Three unrelated upstream moves, all landing in the tests:

QSR overrides sfpsetcc and sfpswap with an extra operand-type argument
the other architectures do not have.  Upstream guards its own calls with
#if __riscv_xtttensixqsr; the shared test bodies need the same, but some
of these calls sit inside backslash-continued macro bodies, so the guard
is a token macro (XTT_IMM_TYPE) rather than an #if around the argument.
23 calls across 14 shared headers.

Upstream now rejects a runtime immediate paired with a null
instruction-buffer pointer -- the synthesised instruction word has to be
written somewhere.  Ten subjects are deliberately dynamic, so they hand
the builtin a buffer instead of nullptr.

invariant-loadi-impostor-bh.C declared its own buffer as unsigned rather
than unsigned long, which the earlier instrn_buffer sweep did not match
because of the name.

+++ gcc/testsuite/g++.target/riscv/tt/tensix/invariant-loadi-impostor-bh.C
@@ -2,7 +2,7 @@
 // { dg-final { scan-tree-dump-not "Hoisted invariant SFPU immediate" "rvtt_invariant" } }
 // { dg-final { scan-assembler-times "SFPLOADI" 2 } }
 
-static volatile unsigned impostor_buffer[1];
+static volatile unsigned long impostor_buffer[1];
 
 void invariant_loadi_impostor ()
 {
